Update git post-checkout hook
When a subommand in the hook returns non zero, the exit code propagates
to the git command that invoked the hook. When we are in the middle of a
rebase, this causes problems and leaves us in a detached head state.
Since running all the various generators for each commit isn't required,
simply shortcircuiting the hook is enough to solve the problem.